+++ CHANGELOG.md
@@ -10,10 +10,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Converted dataset to parquet format for improved performance and faster loading times.
 - Data connects using ibis and duckdb for efficient querying and filtering of the parquet files.
+- Updated CONTRIBUTING.md with M3 collaboration retrospective, and M4 commitments.
+- Added a KPI output that shows the highest and lowest crime rate that is filtered on year, crime category and population (#135). This replaces the population KPI that was there in the previous week's milestone. 
 
 ### Changed
 - Filter section redesigned to be more compact and user-friendly, with filters organized into collapsible sections to save space and improve navigation.
 - Changed Violent Crime Range filter title to "Violent Crime Rate per 100k" to clarify that the filter is based on the crime rate rather than total crime count.
+- Map plot display redesigned to remove scrollbar and bordered outline.
+- Corrected spelling mistakes in the crime category filter
+- Changed the crime rate comparison table to make city-to-city comparisons clearer (#157)
+- Changed the crime rate per 100k KPI title to reflect current crime selection (#155)
+
+
 
 ### Known Issues